Adjust getting the OS token when the running with a different py interpreter
The module to get OS token requires the openstacksdk python module to be
available. This can be an problem if the module is setup in a virtual
environment as in the case of our CLI setup. It is debatable if a virtual
environment is the right way but our current implementation mandates it.

To handle this case, allow the task to use a custom py interpreter to be used
for this task. This requires some jugglery within the ansible tasks to do it two
ways and then consolidate it into a single variable to handle how "register"
action behaves.