Update in 'pre-flight-checks-openstack', main.yml (#288)
If an inventory has only [hypervisors] group defined and does NOT have [cinder] group, existing `when` condition in a 'cinder' related tasks is failing to execute because the current host(`inventory_hostname`) is not found in the `cinder` group. This patch is modifying that `when` condition to first check if the [cinder] group exist in the inventory and then check if the current host is part of it.